pyproject.toml
[tool.poetry]
name = "smart-expense-tracker"
version = "0.1.0"
description = "A smart expense tracker with AI-powered voice input."
authors = ["Hamza khan <hamzakhan127109@gmail.com>"]
readme = "README.md"
packages = [{include = "backend"}]

[tool.poetry.dependencies]
python = ">=3.12"
fastapi = ">=0.116.1"
uvicorn = {extras = ["standard"], version = ">=0.35.0"}
sqlalchemy = ">=2.0.41"
pymysql = ">=1.1.1"
python-dotenv = ">=1.1.1"
alembic = ">=1.16.4"
transformers = "^4.53.2"
torch = "^2.7.1"
torchaudio = "^2.7.1"
sounddevice = "^0.5.2"
scipy = "^1.16.0"
pandas = "^2.3.1"
scikit-learn = "^1.7.0"
langchain-core = "^0.3.69"
python-multipart = "^0.0.20"
pydub = "^0.25.1"
numpy = "^2.3.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"